Goober v10.0
Goober is goated with the sauce now. Man will wait for u to get near then pounce, once discovered, it will hunt the player down, following them as long as they are in the same tile or an adjacent tile. Once the player loses the goober, it will go back to hiding in the tile it it currently in... waiting for its next victim